requirements.txt
flask
scikit-learn
pandas
numpy
tensorflow-cpu==2.15.0
gunicorn
h5py